+++ src/App.jsx
@@ -1,25 +1,14 @@
-import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
 import './App.css'
-
-const toDoList = [
-  {id: 1, title: 'submit assignment'},
-  {id: 2, title: 'submit chapter'},
-  {id: 3, title: 'pass the exam'}
-];
+import TodoList from './toDoList.jsx'
+import AddTodoForm from './AddTodoForm.jsx'
 
 function App() {
-  const [count, setCount] = useState(0)
 
   return (
     <>
       <h1>Todo List</h1>
-      <ul>
-        {toDoList.map((item) => (
-          <li key={item.id}>{item.title}</li>
-        ))}
-      </ul>
+      <TodoList/>
+      <AddTodoForm/>
     </>
   )
 }
